# 默认模块.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "用户登录",
        "deprecated": false,
        "description": "用户登录",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLoginResponse"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "token": "",
                    "userId": 0,
                    "username": "",
                    "role": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "用户注册",
        "deprecated": false,
        "description": "用户注册",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLong"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": 0
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/products/{productId}": {
      "put": {
        "summary": "更新产品",
        "deprecated": false,
        "description": "更新产品",
        "tags": [],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "username",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "FARMER :FARMER\nPURCHASER :PURCHASER\nADMIN :ADMIN",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "FARMER",
                "PURCHASER",
                "ADMIN"
              ]
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FarmerProductRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "summary": "查询产品详情",
        "deprecated": false,
        "description": "查询产品详情",
        "tags": [],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultFarmerProductVO"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "id": 0,
                    "farmerId": 0,
                    "farmName": "",
                    "categoryId": 0,
                    "categoryName": "",
                    "name": "",
                    "spec": "",
                    "unit": "",
                    "price": 0,
                    "minPurchase": 0,
                    "stock": 0,
                    "productionDate": "",
                    "shelfLife": "",
                    "productionMethod": "",
                    "originAreaId": 0,
                    "originAreaName": "",
                    "description": "",
                    "status": "",
                    "images": [
                      ""
                    ],
                    "createTime": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/products/{productId}/on-sale": {
      "put": {
        "summary": "产品上架",
        "deprecated": false,
        "description": "产品上架",
        "tags": [],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "username",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "FARMER :FARMER\nPURCHASER :PURCHASER\nADMIN :ADMIN",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "FARMER",
                "PURCHASER",
                "ADMIN"
              ]
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/products/{productId}/off-sale": {
      "put": {
        "summary": "产品下架",
        "deprecated": false,
        "description": "产品下架",
        "tags": [],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "username",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "FARMER :FARMER\nPURCHASER :PURCHASER\nADMIN :ADMIN",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "FARMER",
                "PURCHASER",
                "ADMIN"
              ]
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/products/my": {
      "get": {
        "summary": "查询我的产品列表",
        "deprecated": false,
        "description": "查询我的产品列表",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "username",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "FARMER :FARMER\nPURCHASER :PURCHASER\nADMIN :ADMIN",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "FARMER",
                "PURCHASER",
                "ADMIN"
              ]
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "pageNum",
            "in": "query",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": true,
            "example": 10,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultIPageFarmerProductVO"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "records": [
                      {
                        "id": 0,
                        "farmerId": 0,
                        "farmName": "",
                        "categoryId": 0,
                        "categoryName": "",
                        "name": "",
                        "spec": "",
                        "unit": "",
                        "price": 0,
                        "minPurchase": 0,
                        "stock": 0,
                        "productionDate": "",
                        "shelfLife": "",
                        "productionMethod": "",
                        "originAreaId": 0,
                        "originAreaName": "",
                        "description": "",
                        "status": "",
                        "images": [
                          ""
                        ],
                        "createTime": ""
                      }
                    ],
                    "total": 0,
                    "size": 0,
                    "current": 0,
                    "orders": [
                      {
                        "column": "",
                        "asc": false
                      }
                    ],
                    "optimizeCountSql": false,
                    "searchCount": false,
                    "optimizeJoinOfCountSql": false,
                    "maxLimit": 0,
                    "countId": "",
                    "pages": 0
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/statistics/product/{productId}": {
      "get": {
        "summary": "获取产品销售统计",
        "deprecated": false,
        "description": "获取产品销售统计",
        "tags": [],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProductSalesStats"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "productName": "",
                    "salesCount": 0,
                    "salesAmount": 0,
                    "reviewCount": 0,
                    "averageRating": 0
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/product-images/product/{productId}": {
      "get": {
        "summary": "按产品ID查询图片列表",
        "deprecated": false,
        "description": "按产品ID查询图片列表",
        "tags": [],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListProductImage"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": [
                    {
                      "id": 0,
                      "productId": 0,
                      "imageUrl": "",
                      "imageType": "",
                      "sort": 0,
                      "createTime": ""
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": ""
          },
          "userId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "username": {
            "type": "string",
            "description": ""
          },
          "role": {
            "type": "string",
            "description": ""
          }
        }
      },
      "ResultLoginResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/LoginResponse",
            "description": ""
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": ""
          },
          "password": {
            "type": "string",
            "description": ""
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "ResultLong": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "type": "integer",
            "description": "",
            "format": "int64"
          }
        }
      },
      "UserRegisterRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": ""
          },
          "password": {
            "type": "string",
            "description": ""
          },
          "role": {
            "type": "string",
            "description": "",
            "enum": [
              "FARMER",
              "PURCHASER",
              "ADMIN"
            ]
          },
          "contactPerson": {
            "type": "string",
            "description": ""
          },
          "contactPhone": {
            "type": "string",
            "description": "",
            "pattern": "^1[3-9]\\d{9}$"
          },
          "contactEmail": {
            "type": "string",
            "description": "",
            "format": "email"
          }
        },
        "required": [
          "username",
          "password",
          "role",
          "contactPerson",
          "contactPhone"
        ]
      },
...
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "description": "",
            "type": "null"
          }
        }
      },
...
      "FarmerProductRequest": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": ""
          },
          "spec": {
            "type": "string",
            "description": ""
          },
          "unit": {
            "type": "string",
            "description": ""
          },
          "price": {
            "type": "number",
            "description": "",
            "minimum": 0.01
          },
          "minPurchase": {
            "type": "integer",
            "description": "",
            "minimum": 1
          },
          "stock": {
            "type": "integer",
            "description": "",
            "minimum": 1
          },
          "productionDate": {
            "type": "string",
            "description": ""
          },
          "shelfLife": {
            "type": "string",
            "description": ""
          },
          "productionMethod": {
            "type": "string",
            "description": ""
          },
          "originAreaId": {
            "type": "integer",
            "description": ""
          },
          "description": {
            "type": "string",
            "description": ""
          }
        },
        "required": [
          "categoryId",
          "name",
          "spec",
          "unit",
          "price",
          "minPurchase",
          "stock",
          "originAreaId"
        ]
      },
...
      "FarmerProductVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "farmerId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "farmName": {
            "type": "string",
            "description": "关联农场名称"
          },
          "categoryId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "categoryName": {
            "type": "string",
            "description": "关联分类名称"
          },
          "name": {
            "type": "string",
            "description": ""
          },
          "spec": {
            "type": "string",
            "description": ""
          },
          "unit": {
            "type": "string",
            "description": ""
          },
          "price": {
            "type": "number",
            "description": ""
          },
          "minPurchase": {
            "type": "integer",
            "description": ""
          },
          "stock": {
            "type": "integer",
            "description": ""
          },
          "productionDate": {
            "type": "string",
            "description": ""
          },
          "shelfLife": {
            "type": "string",
            "description": ""
          },
          "productionMethod": {
            "type": "string",
            "description": ""
          },
          "originAreaId": {
            "type": "integer",
            "description": ""
          },
          "originAreaName": {
            "type": "string",
            "description": "关联产地名称"
          },
          "description": {
            "type": "string",
            "description": ""
          },
          "status": {
            "type": "string",
            "description": "",
            "enum": [
              "ON_SALE",
              "OFF_SALE"
            ]
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "产品图片列表"
          },
          "createTime": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "IPageFarmerProductVO": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FarmerProductVO",
              "description": "农户产品视图对象"
            },
            "description": "查询数据列表",
            "default": "Collections.emptyList()"
          },
          "total": {
            "type": "integer",
            "description": "总数",
            "format": "int64",
            "default": 0
          },
          "size": {
            "type": "integer",
            "description": "每页显示条数，默认 10",
            "format": "int64",
            "default": 10
          },
          "current": {
            "type": "integer",
            "description": "当前页",
            "format": "int64",
            "default": 1
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem",
              "description": "com.baomidou.mybatisplus.core.metadata.OrderItem"
            },
            "description": "排序字段信息",
            "default": "new ArrayList<>()"
          },
          "optimizeCountSql": {
            "type": "boolean",
            "description": "自动优化 COUNT SQL",
            "default": true
          },
          "searchCount": {
            "type": "boolean",
            "description": "是否进行 count 查询",
            "default": true
          },
          "optimizeJoinOfCountSql": {
            "type": "boolean",
            "description": "{@link #optimizeJoinOfCountSql()}",
            "default": true
          },
          "maxLimit": {
            "type": "integer",
            "description": "单页分页条数限制",
            "format": "int64"
          },
          "countId": {
            "type": "string",
            "description": "countId"
          },
          "pages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ResultIPageFarmerProductVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/IPageFarmerProductVO",
            "description": ""
          }
        }
      },
...
      "ResultFarmerProductVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/FarmerProductVO",
            "description": ""
          }
        }
      },
...
      "ProductSalesStats": {
        "type": "object",
        "properties": {
          "productName": {
            "type": "string",
            "description": "产品名称"
          },
          "salesCount": {
            "type": "integer",
            "description": "销售数量"
          },
          "salesAmount": {
            "type": "number",
            "description": "销售总额"
          },
          "reviewCount": {
            "type": "integer",
            "description": "评价数"
          },
          "averageRating": {
            "type": "number",
            "description": "平均评分"
          }
        }
      },
...
      "ResultProductSalesStats": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/ProductSalesStats",
            "description": ""
          }
        }
      },
...
      "ProductImage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "productId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "imageUrl": {
            "type": "string",
            "description": ""
          },
          "imageType": {
            "type": "string",
            "description": "",
            "enum": [
              "COVER",
              "PRODUCTION",
              "DETAIL"
            ]
          },
          "sort": {
            "type": "integer",
            "description": ""
          },
          "createTime": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "ResultListProductImage": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImage",
              "description": "产品图片实体类"
            },
            "description": ""
          }
        }
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}